[#180984438] - Improve documentation and add additional tests (#21)
* [#180969915] - Add tests and Fix module vulnerabilities (#15)

* Fix vulneabilities in some modules (#14)

* [#180970863] - Add integrations (#16)

Improvement

* Add redefined documentation

Using swagger components. The errors are now easy to
maintain and retain similar form across the board

* [#180971362] - Refined documentation (#19)

* Fix vulneabilities in some modules (#14)

* [#180970863] - Add integrations (#16)

Improvement

* Add redefined documentation

Using swagger components. The errors are now easy to
maintain and retain similar form across the board

* [#180969915]   Testing the blog routes (#20)

* Add blog test

The additional tests ensures that not only data,
but permissions are kept across the application

* Add blog test

The additional tests ensures that not only data,
but permissions are kept across the application

* [#180969915]   Testing the blog routes (#20)

* Add blog test

The additional tests ensures that not only data,
but permissions are kept across the application

* Add blog test

The additional tests ensures that not only data,
but permissions are kept across the application

* Fix error of multiple returns and logout view

* Fix error of multiple returns and logout view

* Fix user deletion